[refactor] : TransactionList View 분리 및 수정
